refactor: remove `js-sha1` package to reduce bundle size
Removes the `js-sha1` package and replaces it with the browser `SubtleCrypto` API.

We add a `cryptoSubtleFeatureTest` to test if we should use this API. If we're able to use the API then we perform the
breached checked as normal.

If the feature test fails either because the user is using Internet Explorer or the Web Crypto API is not supported, then
we skip the breached check on the client side, and assume a valid password has been entered if between 8 - 72 characters.

The breached password check happens on the server side anyway, so if the user has entered a breached password on the
client, we'll pick it up anyway on the server side and ask the user to change it should they be using an unsupported
browser.

Tested support using Browserstack in IE11 and older chrome versions.