refactor: remove `query-string` package to reduce bundle size
Removes the `query-string` package and replaces it with the browser `URLSearchParams` API.

Polyfill `URLSearchParams` for browsers that don't support the API.

Add additional methods and update tests to make sure this functionality remains identitcal
to how we use it.

This saves a small amount in our bundle size, as well as removes the need for the
dependency directly.