fix: Add `The Guardian` to the `FromEmailAddress` in email sending
Emails from gatway were being sent from the

registration-reply@theguardian.com

email address, but appeared in the inbox as just `registration-reply`
with no relation to who sent the message.

By using `The Guardian <registration-reply@theguardian.com>` as the
`FromEmailAddress`, it will appear in the inbox as an email from `The
Guardian` making it easier to see and find it.