fix(useInputValidityState): only transition validity state when input is not empty
An issue occurs with the built in password manager in safari where it
aggresively tries to fill in the email and password field should a
password exist for that email. When the users blurs the input field
without entering an email/password the `Please enter your email.` error
is shown, despite the user not wanting to take this action.

We add a check to the `onBlur` method so that it won't attempt to change
the validation state when clicked away from the input field if it is
empty.