refactor(GU_SO cookie): set and clear sign out cookie from gateway
In preparation for the Okta migration for web applications, we want to be able to tell when a user has signed out within the last 24 hours on
a given device.

We do this by reusing the existing `GU_SO` cookie (signout cookie) which we previously got from Identity API which was set when a user signed
out. This cookie has a value which is the unix epoch time in seconds when the user signed out, and an expiry of 24 hours. It's also set as a
non-HTTPOnly cookie so that it can be read in javascript applications.

This allows any applications when reading this cookie to assume that the user is currently signed out, and clear any tokens/user data
required if it comes across this cookie.