feat: react types to v18
Updates `@types/react` and `@types/react-dom` to v18 in prep for the department wide upgrade to React 18.

Since we're using `preact/compat` the only things we should need to change are the type definitions, as everything else
should behave as it does in React 17.

The main type changes we needed to make were to explicitly define `children` using `PropsWithChildren` in some of our
components.

The only other issue resided with our use of `mjml-react` which was showing type errors.

On further investigation, the `mjml-react` project has been unmaintained for a while:
https://github.com/wix-incubator/mjml-react

It recommended that we switch to a fork called `@faire/mjml-react`: https://github.com/Faire/mjml-react

This fixed all the type issues, and other problems we had with `mjml-react` previously.

Overall it was a pleasant enough upgrade to React 18 types!