fix(loginMiddlewareOauth): check for valid session inside `loginMiddlewareOAuth` to avoid auth code flow nesting
Raphael Kabo <raphael.kabo@theguardian.com> noticed that when attempting to navigate to a page using `loginMiddlewareOAuth` while signed out, it would redirect to a login page, and then when attempting to log in it would show the `signed in as` page with a `Something went wrong` error
rather than the page the user was trying to access. The logs showed an `{"level":"error","message":"Missing auth state cookie on OAuth Callback!","request_id":"xxx"}` error.

Investigating this, I found that this occured because 2 OAuth Authorization Code flows were attempted at the same time, one initiated by the other. However both of these flows rely on the `GU_oidc_auth_state` cookie, and the inner auth code flow would overwrite the
`GU_oidc_auth_state` cookie set by the other flow. So after the inner flow completed, it would attempt to complete the outer flow, but is not able to because the `GU_oidc_auth_state` cookie had been overwritten or deleted, hence an error is thrown!

This occurs because the `loginMiddlewareOAuth` didn't check if the user had a valid session or not, just if it had valid OAuth tokens. So the `loginMiddlewareOAuth` would initiate an OAuth Auth Code flow to get new tokens. If the user was signed in this would complete successfully
without needed another code flow, but if the user was signed out they would have to authenticate first, which would cause another Authorization Code flow to happen, as authentication needs its own code flow in order to complete!

This commit fixes this issue by checking if the `sid` cookie exists and the user has a valid session in the `loginMiddlewareOAuth` middleware itself. If the user does have a valid session then the rest of the middleware behaves as it did before. If the user does not have a valid
session we now redirect the user to the signin page, with a `returnUrl` query parameter which will take them back to the page they were trying to access once they do authenticate.

This effectively means that the 2 Authorization code flows happen sequentially and independantly of each other. First the "authentication" one will happen when the user authenticates, then the "application" one to get tokens for the application itself, rather than the "application"
one initiaing the "authentication" one inside of itself.