feat(register): add `getRegistrationPlatform` method and correct `registrationPlatform` when creating user
Currently the `registrationPlatform` parameter on the user object within Otka is always set to `identity-gateway`, even if the user registers from a native app or another web application like mma (manage).

This means we lose the information about where the user registered from, which is useful for tracking purposes.

We were already providing the `appClientId` to the registration method when creating a user. We can use this ID to call the `getApp` function in order to retrieve the app name from Okta, and use this as the registration platform. This applies to all applications which use OAuth currently to sign in and register.

We also change the default from `identity-gateway` to `profile`, as `profile` is the name of the Gateway application (this project) within Okta.