Add /unsubscribe-all endpoint (#2731)
Adds a new /unsubscribe-all endpoint, to be used in the List-Unsubscribe header
of emails.

According to [the spec] this must be a POST endpoint. By default CSRF protection
is enabled on all POSTs, but that's not appropriate in this case as the request
is made from an external location. So I've had to modify the middleware to
allow POSTs to this endpoint with no CSRF token.

See also guardian/identity#2514.

[the spec]: https://datatracker.ietf.org/doc/html/rfc8058