fix(passcodes): replace `/register/code/expired` with `/welcome/expired`
If a user takes too long to type a passcode, they should be redirected to an expired page.

The idea initially was to make a new `/register/code/expired` page. This was never implemented in the original PR, so if a user ended up in this scenario they would see a 404 page.

At first we thought we should create a new page under this url, but since this has the exact same functionality as `/welcome/expired` we should just use that one, which is what this commit does.