refactor(EmailSent): create new `PasscodeEmailSent` page if using passcodes
The display logic in the `EmailSent` page component was getting super messy and difficult to reason about, more so when passcodes were added in.

It is a sensible decision to make a new EmailSent page component for when we're using passcodes. So we create a `PasscodeEmailSent` component to handle this behaviour.