Move link expired message to footer
We believe there is a risk that users might confuse the two links in the body of the password reset email: the main call to action that enables the user to reset their password, and the recently-added instructions for resending if the link has expired. This change aims to remedy any user confusion by moving the instructions to the footer.

Signed-off-by: Nick Hart <126492147+nickhartyyy@users.noreply.github.com>