refactor(idx): add helper `ExtractLiteralRemediationNames` type
This is used to extract all possible remediation `name` values into a template literal type containing all the possible values except `string`.

For example:

```ts
// Type to extract all the remediation names from the introspect response
type IntrospectRemediationNames = ExtractLiteralRemediationNames<
	IntrospectResponse['remediation']['value'][number]
>;
```

Will evaluate to:

```ts
type IntrospectRemediationNames = "redirect-idp" | "select-enroll-profile" | "identify"
```

Which is helpful as it means we can avoid using the `string` type when validating the remediation

For example in the `validateIntrospectRemediation` which now uses this type.