refactor(idx): make `IntrospectResponse` remediation a union of all other remediations
Since the `introspect` call can be called at any poin in the IDX flow to check the current state of the flow, the response returned by the introspect call could be the same as any other IDX API call.

This means that the remediation object could be any remediation object from the other IDX API calls.

This commit updates the remediation object inside the `IntrospectResponse` to be a union of all the other remediation objects from the other IDX API calls.