refactor(idx): create `findAuthenticatorId` helper method
To extract a given authenticators id from within the remediation object is a bit of a mess.

Since this functionality will be needed quite often, I've moved this to it's own method.

This method handles all the current possible scenarios where we'll need to find authenticatior ids.

Currently in the create account flow it's using the `select-authenticator-enroll` remediation to find the authenticator id for the `password` authenticator.

When we start doing passwordless we'll need more options here. In this commit I've also added the two that are required for password reset using passcodes, specifically `select-authenticator-authenticate` and `authenticator-verification-data`.

Since the shape to find the id can sometimes be different, this also takes into account the possible schema shapes too, so we can hid the complexity from the usages.