refactor(idx): remove `idx` cookie from `idxFetch` methods
From testing we don't actually need to get the idx cookie from the completion response, it will be automatically set when we redirect the user to the `/login/token/redirect` endpoint which sets the correct global session (idx) cookie.

By removing this code, we can make the `idxFetch` methods easier to reason about.