feat(idx): add `submitPassword` method
This is similar to the `submitPasscode` method, but takes a password instead.

It submit a password to Okta to answer the password challenge using the `challenge/answer` endpoint and return the response.

Validation can be enabled as needed through the params, `validatePasswordLength` and `validateBreachedPassword` (both default to false) where if validation fails then the method will throw.