refactor(dangerouslySetPlaceholderPassword): add param to return placeholder password if true
In some cases, e.g. for ACTIVE users with only the "password" authenticator, we actually need to set a placeholder password for these users that is then used to authenticate the user in order for them to be able to verify their email. To do this we need to return the placeholder password so we can use it in the Okta IDX authentication process.