feat(idx): add `validateChallengeRemediation` method for the `challengeResponse`
A function that validates that the challenge response contains a remediation with the given name and authenticator.

This is slightly more complicated compared to the other `validateRemediation` methods, as we don't just want to check if the response has a specific remediation name, we also want to check if it has a specific authenticator, and possibly also a given resend/recover method.