refactor(idx): remove `idxFetchCompletion` and update `setPasswordAndRedirect` to use `submitPassword` instead
In Okta IDX API, it is difficult to determine the exact scenarios where we get a `CompleteLoginResponse`, and thus when to use the `idxFetchCompletion` method vs the `idxFetch` method.

We previously refactored `challenge/answer` to be either `ChallengeAnswerResponse` or `CompleteLoginResponse`, and added the `isChallengeAnswerCompleteLoginResponse` and `isChallengeAnswerResponse` type guards which should be used to check which response we recieved.

We also made a new `submitPassword` function, which abstracts away everything to do with submiting the password to the Okta API and any additional validation.

We can then update the `setPasswordAndRedirect` method to use `submitPassword` instead of the `idxFetchCompletion` method, and use the new `getLoginRedirectUrl` to determine where the user needs to go to get a global session set after authenticating.