feat(idx): passcode password reset for non-ACTIVE users
For users in a non-ACTIVE state, we should first get them into one of the ACTIVE states in order to send them an email which allows them to reset their password with a passcode.

The best way to do this is to first deactivate the user, which works on all user states and puts them into the DEPROVISIONED state.

Then we can activate the user, which will put them into the PROVISIONED state and return us a recovery token.

We then use the recovery token to set a placeholder password for the user, which transitions them into the ACTIVE state and then we can call `changePasswordEmailIdx` method again to send the user a passcode as they'll be in one of the ACTIVE states