refactor(PasscodeEmailSent): add `textType` parameter to determine text on the page
Rather than use the route to determine which text to use on the `PasscodeEmailSent` page component, we use a prop instead.

The `textType` can be either `verification`, `security`, or `generic`.

`generic` - This is the default text for the page, currently used by the `/reset-password/email-sent` page, and in the future will also be used for passcodes via sign in
`verification` - This is used for the "create account"/"registration" journey, i.e `/register/email-sent`, and any future passcodes sent via that journey
`security` - Used by the sign in flow where users need to re-verify their accounts or reset their passwords, slightly different text from the above saying "For security reasons we need you to change your password.". Used by `/signin/email-sent`.