fix(create-account): add check for `enroll-authenticator` and email authenticator
In https://github.com/guardian/gateway/pull/3127, we attemtped to make the setting of passwords optional during the account creation flow.

As part of this we had to update the "password" authenticator in Okta from "required" to "optional", this was done in https://github.com/guardian/identity-platform/pull/806

However after this was launched we quickly noticed that new users were getting two verification codes sent immediatley one after the other when they attemoted to create an account.

This would mean that one of the codes was valid while the other one wasn't!

We realised that this was down to how the IDX API had changed when we updated the "password" authenticator in Okta from "required" to "optional".

When the "password" authenticator is "required", the `/idp/idx/enroll/new` endpoint returns a remediation similar to the following:

```
{
	...
	"remediation": {
		"type": "array",
		"value": [
		  {
			...
			"name": "select-authenticator-enroll",
			...
			"value": [
			  {
				"name": "authenticator",
				"type": "object",
				"options": [
				  {
					"label": "Email",
					"value": {
					  "form": {
						"value": [
						  {
							"name": "id",
							"required": true,
							"value": "...",
							"mutable": false
						  },
						  {
							"name": "methodType",
							"required": false,
							"value": "email",
							"mutable": false
						  }
						]
					  }
					},
					"relatesTo": "$.authenticators.value[0]"
				  },
				  {
					"label": "Password",
					"value": {
					  "form": {
						"value": [
						  {
							"name": "id",
							"required": true,
							"value": "...",
							"mutable": false
						  },
						  {
							"name": "methodType",
							"required": false,
							"value": "password",
							"mutable": false
						  }
						]
					  }
					},
					"relatesTo": "$.authenticators.value[1]"
				  }
				]
			  },
			  {
				"name": "stateHandle",
				"required": true,
				"value": "...",
				"visible": false,
				"mutable": false
			  }
			],
			...
		  }
		]
	  }
	  ...
}
```

Where we only have a `select-authenticator-enroll` object, with us having to choose which authenticator we want to set (we chose "email" in order to verify the email first before setting the "password" authenticator after).

However after updating the "password" authenticator to optional in Okta, the `/idp/idx/enroll/new` endpoint now returns something like the following:

```
{
	...
	"remediation": {
		"value": [
		  {
			...
			"name": "enroll-authenticator",
			...
			"value": [
			  {
				"name": "credentials",
				"type": "object",
				"form": {
				  "value": [
					{
					  "name": "passcode",
					  "label": "Enter code"
					}
				  ]
				},
				"required": true
			  },
			  {
				"name": "stateHandle",
				"required": true,
				"value": "...",
				"visible": false,
				"mutable": false
			  }
			],
			"accepts": "application/json; okta-version=1.0.0"
		  },
		  {
			"name": "select-authenticator-enroll",
			...
			"value": [
			  {
				"name": "authenticator",
				"type": "object",
				"options": [
				  {
					"label": "Email",
					"value": {
					  "form": {
						"value": [
						  {
							"name": "id",
							"required": true,
							"value": "...",
							"mutable": false
						  },
						  {
							"name": "methodType",
							"required": false,
							"value": "email",
							"mutable": false
						  }
						]
					  }
					},
					"relatesTo": "$.authenticators.value[0]"
				  }
				]
			  },
			  {
				"name": "stateHandle",
				"required": true,
				"value": "...",
				"visible": false,
				"mutable": false
			  }
			],
			"accepts": "application/json; okta-version=1.0.0"
		  }
		]
	  },
	  "currentAuthenticator": {
		"value": {
		  "resend": {
			...
			"name": "resend",
			...
		  },
		  "type": "email",
		  ...
		}
	  },
	  ...
}
```

Where we have both an `enroll-authenticator` and `select-authenticator-enroll` remediation, plus a `currentAuthenticator` object.

This tells us that an email has automatically been sent to the reader, and we shouldn't have to check `select-authenticator-enroll`.

Our mistake was that we were only checking for the existence of the `select-authenticator-enroll` remediation, rather than checking for the `enroll-authenticator` and that the `currentAuthenticator` was "email".

So what was happening was that Okta was automatically sending the user an email after we called `/idp/idx/enroll/new` with a passcode when the "password" authenticator was set to "optional", but we were still checking for the `select-authenticator-enroll` remediation, and essentially sending the user a 2nd passcode straight away by calling the `/idp/idx/credential/enroll` endpoint inside `select-authenticator-enroll`!

To fix this we have to check wether we have `enroll-authenticator` and that the `currentAuthenticator` was "email". In this scenario we should skip checking the `select-authenticator-enroll` remediation, which would be the case when "password" is optional.

If they don't exist, and `select-authenticator-enroll` still does, then we should manually send the user an email by using the  `select-authenticator-enroll` remediation, which would be the case when "password" is required.