Add user-benefits-api (#3124)
<!-- See
https://github.com/guardian/recommendations/blob/main/pull-requests.md
for recommendations on raising and reviewing pull requests. -->

## What does this change?

This PR adds functionality to set the required product benefit cookies
upon user sign-in and session refresh.

It retrieves the user’s product benefits by making a request to the
`/benefits/me` endpoint of the user-benefits-api. The response returns
an array of benefit identifiers associated with the user.

Based on the returned benefits, the following cookies are conditionally
set: `gu_user_benefits_expiry`, `gu_hide_support_messaging`,
`gu_allow_reject_all`, `GU_AF1`

This PR builds on prior work in Dotcom, referencing:

https://github.com/guardian/dotcom-rendering/blob/a0eff55ee7fedf08b785557fceaa0f4f7265e0df/dotcom-rendering/src/client/userFeatures/user-features.ts

https://github.com/guardian/support-service-lambdas/blob/main/modules/product-benefits/src/schemas.ts

This change introduces support for the new user-benefits-api, with the
aim of eventually replacing reliance on the members-data-api.


https://docs.google.com/document/d/19ExTKgMhCJbrSLD9GFTeNKy3ArjoS2GR8wVc2ONbNXo/edit?tab=t.0

<!-- A PR should have enough detail to be understandable far in the
future. e.g what is the problem/why is the change needed, how does it
solve it and any questions or points of discussion. Prefer copying
information from a Trello card over linking to it; the card may not
always exist and reviewers may not have access to the board. -->

## How to test

<!-- Provide instructions to help others verify the change. This could
take the form of "On PROD, do X and witness Y. On this branch, do X and
witness Z. " -->

## How can we measure success?

<!-- Do you expect errors to decrease? Do you expect user journeys to be
simplified? What can be used to prove this? A filtered view of logs or
analytics, etc? -->