Remove Trump On Trial and add Tech-Scape newsletter in US (#3234)
<!-- See
https://github.com/guardian/recommendations/blob/main/pull-requests.md
for recommendations on raising and reviewing pull requests. -->

## What does this change?
This PR removes the Trump on Trial newsletter from the US and adds the
Tech-Scape newsletter, as requested by the US team.

The difference in screenshots reflects the difference in the newsletter
setup for CODE and PROD. The the-stakes-us-election-edition newsletter
has a different in CODE and PROD hence it's not retrieved in the CODE
environment.

https://trello.com/c/QQ7Yth1d


| Before  | After |
| ------------- | ------------- |
| <img width="1512" height="859" alt="Screenshot 2025-07-21 at 11 18 41"
src="https://github.com/user-attachments/assets/7597c7ca-5aa5-462f-8080-baee8649a327"
/>| <img width="1512" height="949" alt="Screenshot 2025-07-22 at 13 25
07"
src="https://github.com/user-attachments/assets/aca371e4-d743-4de2-a069-8dfe69cfeb8e"
/>|



<!-- A PR should have enough detail to be understandable far in the
future. e.g what is the problem/why is the change needed, how does it
solve it and any questions or points of discussion. Prefer copying
information from a Trello card over linking to it; the card may not
always exist and reviewers may not have access to the board. -->

## How to test

<!-- Provide instructions to help others verify the change. This could
take the form of "On PROD, do X and witness Y. On this branch, do X and
witness Z. " -->
This has been tested locally and deploying to CODE. To ensure it worked,
I made the techscape live in the newsletter CODE.

## Images

<!-- Usually only applicable to UI changes, what did it look like before
and what will it look like after? -->

![Uploading Screenshot 2025-07-22 at 14.05.53.png…]()