Show /welcome/review page for Feast app. (#3235)
<!-- See
https://github.com/guardian/recommendations/blob/main/pull-requests.md
for recommendations on raising and reviewing pull requests. -->

## What does this change?
This PR continues the work of this
https://github.com/guardian/gateway/pull/3207

This is onboarding the feast app to use the /welcome/review page.

When creating a new account via any Client thats not
[www.theguardian.com](http://www.theguardian.com/) we skip the ["Welcome
to The Guardian"
page](https://github.com/user-attachments/assets/2b69d322-4752-4b99-8417-322020918f2b)
which asks the users if they consent to their signed in data being used
for marketing content.

Since the user isn't shown this page we expected them to be opted out of
this consent, however, as we [treat this consent as an "opt out"
consent](https://github.com/guardian/identity/blob/0580155e9eac0fe52ef4772fa1ae72d91155f258/identity-model/src/main/scala/com/gu/identity/model/Consent.scala#L211)
unfortunately it means if no action is taken then the user is
automatically opted in.

We've been given permission to add the "Welcome to The Guardian"
onboarding page to the feast flow so that we can ask users for consent.

<!-- A PR should have enough detail to be understandable far in the
future. e.g what is the problem/why is the change needed, how does it
solve it and any questions or points of discussion. Prefer copying
information from a Trello card over linking to it; the card may not
always exist and reviewers may not have access to the board. -->