Remove `.only` exclusivity modifier applied to test. (#3272)
<!-- See
https://github.com/guardian/recommendations/blob/main/pull-requests.md
for recommendations on raising and reviewing pull requests. -->

## What does this change?
Remove exclusivity feature (.only) applied to tests in
cypress/integration/ete/sign_in.1.cy.ts and
cypress/integration/mocked/okta_sign_in.6.cy.ts.
 
<!-- A PR should have enough detail to be understandable far in the
future. e.g what is the problem/why is the change needed, how does it
solve it and any questions or points of discussion. Prefer copying
information from a Trello card over linking to it; the card may not
always exist and reviewers may not have access to the board. -->

## How to test
Run the mocked and e2e cypress tests.
<!-- Provide instructions to help others verify the change. This could
take the form of "On PROD, do X and witness Y. On this branch, do X and
witness Z. " -->